tests/intel/xe_eudebug_online: single-step support for XE3P
Adding support for single-step tests for platforms with
per-context-debugging support.
There is no long bitmask with threads that can be manipulated.
Instead the userspace -> memory -> SIP interaction shall be performed.
For every test dedicated SIP is required.

This test stops all threads on a breakpoint.
Then single thread is resumed with single-stepping on twice.
With every resume following checks are performed:
- if there are no other exceptions than breakpoint (for all threads)
- if AIP of each theads is right (only for resumed thread it can change).

After all threads are exercised shader is resumed without
single-steppin on to exit.

Signed-off-by: Maciej Patelczyk <maciej.patelczyk@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>